+++ extensions/ql-vscode/src/debugger/debug-configuration.ts
@@ -4,7 +4,8 @@ import {
   DebugConfigurationProvider,
   WorkspaceFolder,
 } from "vscode";
-import { getOnDiskWorkspaceFolders, showAndLogErrorMessage } from "../helpers";
+import { showAndLogErrorMessage } from "../helpers";
+import { getOnDiskWorkspaceFolders } from "../common/vscode/workspace-folders";
 import { LocalQueries } from "../local-queries";
 import { getQuickEvalContext, validateQueryPath } from "../run-queries-shared";
 import * as CodeQLProtocol from "./debug-protocol";
