+++ extensions/ql-vscode/src/helpers.ts
@@ -1,7 +1,7 @@
 import { ensureDirSync, pathExists, ensureDir, writeFile } from "fs-extra";
-import { join, dirname } from "path";
+import { join } from "path";
 import { dirSync } from "tmp-promise";
-import { Uri, window as Window, workspace, env, WorkspaceFolder } from "vscode";
+import { Uri, window as Window, workspace, env } from "vscode";
 import { CodeQLCliServer } from "./codeql-cli/cli";
 import { UserCancellationException } from "./common/vscode/progress";
 import { extLogger, OutputChannelLogger } from "./common";
@@ -11,6 +11,7 @@ import { RedactableError } from "./pure/errors";
 import { isQueryLanguage, QueryLanguage } from "./common/query-language";
 import { isCodespacesTemplate } from "./config";
 import { AppCommandManager } from "./common/commands";
+import { getOnDiskWorkspaceFolders } from "./common/vscode/workspace-folders";
 
 // Shared temporary folder for the extension.
 export const tmpDir = dirSync({
@@ -232,13 +233,6 @@ export async function showInformationMessageWithAction(
   return chosenItem === actionItem;
 }
 
-/** Returns true if the specified workspace folder is on the file system. */
-export function isWorkspaceFolderOnDisk(
-  workspaceFolder: WorkspaceFolder,
-): boolean {
-  return workspaceFolder.uri.scheme === "file";
-}
-
 /**
  * Opens a modal dialog for the user to make a choice between yes/no/never be asked again.
  *
@@ -277,26 +271,6 @@ export async function showNeverAskAgainDialog(
   );
 
   return chosenItem?.title;
-}
-
-/** Gets all active workspace folders that are on the filesystem. */
-export function getOnDiskWorkspaceFoldersObjects() {
-  const workspaceFolders = workspace.workspaceFolders ?? [];
-  return workspaceFolders.filter(isWorkspaceFolderOnDisk);
-}
-
-/** Gets all active workspace folders that are on the filesystem. */
-export function getOnDiskWorkspaceFolders() {
-  return getOnDiskWorkspaceFoldersObjects().map((folder) => folder.uri.fsPath);
-}
-
-/** Check if folder is already present in workspace */
-export function isFolderAlreadyInWorkspace(folderName: string) {
-  const workspaceFolders = workspace.workspaceFolders || [];
-
-  return !!workspaceFolders.find(
-    (workspaceFolder) => workspaceFolder.name === folderName,
-  );
 }
 
 /** Check if the current workspace is the CodeTour and open the workspace folder.
@@ -447,40 +421,4 @@ export async function createTimestampFile(storagePath: string) {
   const timestampPath = join(storagePath, "timestamp");
   await ensureDir(storagePath);
   await writeFile(timestampPath, Date.now().toString(), "utf8");
-}
-
-/**
- * Returns the path of the first folder in the workspace.
- * This is used to decide where to create skeleton QL packs.
- *
- * If the first folder is a QL pack, then the parent folder is returned.
- * This is because the vscode-codeql-starter repo contains a ql pack in
- * the first folder.
- *
- * This is a temporary workaround until we can retire the
- * vscode-codeql-starter repo.
- */
-
-export function getFirstWorkspaceFolder() {
-  const workspaceFolders = getOnDiskWorkspaceFolders();
-
-  if (!workspaceFolders || workspaceFolders.length === 0) {
-    throw new Error("No workspace folders found");
-  }
-
-  const firstFolderFsPath = workspaceFolders[0];
-
-  // For the vscode-codeql-starter repo, the first folder will be a ql pack
-  // so we need to get the parent folder
-  if (
-    firstFolderFsPath.includes(
-      join("vscode-codeql-starter", "codeql-custom-queries"),
-    )
-  ) {
-    // return the parent folder
-    return dirname(firstFolderFsPath);
-  } else {
-    // if the first folder is not a ql pack, then we are in a normal workspace
-    return firstFolderFsPath;
-  }
 }
