+++ extensions/ql-vscode/src/queries-panel/query-discovery.ts
@@ -5,7 +5,7 @@ import { Event, RelativePattern, Uri, WorkspaceFolder } from "vscode";
 import { MultiFileSystemWatcher } from "../common/vscode/multi-file-system-watcher";
 import { App } from "../common/app";
 import { FileTreeDirectory, FileTreeLeaf } from "../common/file-tree-nodes";
-import { getOnDiskWorkspaceFoldersObjects } from "../helpers";
+import { getOnDiskWorkspaceFoldersObjects } from "../common/vscode/workspace-folders";
 import { AppEventEmitter } from "../common/events";
 import { QueryDiscoverer } from "./query-tree-data-provider";
 import { extLogger } from "../common";
